fix: inner solver tolerances derive from atol/rtol under fixed step control
atol and rtol live on BaseStepControllerConfig, so every step
controller (fixed included) accepts, stores, and updates them. The
inner Newton/Krylov tolerance derivation applies whenever the
algorithm is implicit, regardless of controller type; explicitly set
inner tolerances still win. The fixed controller substituted by
check_compatibility for errorless algorithms carries the replaced
controller's atol/rtol, so backwards_euler solves keep the user's
request.
